+++ launch/action_demo.py
@@ -1,0 +1,22 @@
+from launch import LaunchDescription
+from launch_ros.actions import Node
+
+
+def generate_launch_description():
+    return LaunchDescription(
+        [
+            Node(
+                package="ros_tcp_endpoint",
+                executable="default_server_endpoint",
+                output="screen",
+                emulate_tty=True,
+                parameters=[{"ROS_IP": "0.0.0.0"}, {"ROS_TCP_PORT": 10000}],
+            ),
+            Node(
+                package="ros_tcp_endpoint",
+                executable="fibonacci_action_server",
+                output="screen",
+                emulate_tty=True,
+            ),
+        ]
+    )